Moved ui to src
