+++ src/ui/Button/__tests__/Button.tests.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Button from '..'
+import Button from '../../Button'
 import renderer from 'react-test-renderer'
 
 describe('Button', () => {
