+++ src/ui/FormattedCarousel/FormattedCarousel.js
