+++ src/ui/FormattedCarousel/__tests__/FormattedCarousel.tests.js
