+++ src/ui/FormattedCarousel/index.js
