+++ src/ui/GroupBlock/__tests__/GroupBlock.tests.js
