+++ src/ui/inputs/inputIcon.js
